move tests to a separate folder
